move metric_context.py and metric_monitor.py into common module
